refactor: 클래스 분리 및 변수명 변경
Validator와 RaceGame로 기능 별 클래스 분리
